Chore: remove models.py and split its contents to api files
