+++ client/cloud_credentials.go
@@ -4,6 +4,79 @@ import (
 	"fmt"
 )
 
+type AwsCredentialsType string
+type GcpCredentialsType string
+type AzureCredentialsType string
+
+type Credentials struct {
+	Id             string `json:"id"`
+	Name           string `json:"name"`
+	OrganizationId string `json:"organizationId"`
+	Type           string `json:"type"`
+}
+
+type AzureCredentialsCreatePayload struct {
+	Name           string                       `json:"name"`
+	OrganizationId string                       `json:"organizationId"`
+	Type           AzureCredentialsType         `json:"type"`
+	Value          AzureCredentialsValuePayload `json:"value"`
+}
+
+type AzureCredentialsValuePayload struct {
+	ClientId       string `json:"clientId"`
+	ClientSecret   string `json:"clientSecret"`
+	SubscriptionId string `json:"subscriptionId"`
+	TenantId       string `json:"tenantId"`
+}
+
+type AwsCredentialsCreatePayload struct {
+	Name           string                     `json:"name"`
+	OrganizationId string                     `json:"organizationId"`
+	Type           AwsCredentialsType         `json:"type"`
+	Value          AwsCredentialsValuePayload `json:"value"`
+}
+
+type AwsCredentialsValuePayload struct {
+	RoleArn         string `json:"roleArn"`
+	ExternalId      string `json:"externalId"`
+	AccessKeyId     string `json:"accessKeyId"`
+	SecretAccessKey string `json:"secretAccessKey"`
+}
+
+type GoogleCostCredentialsCreatePayload struct {
+	Name           string                            `json:"name"`
+	OrganizationId string                            `json:"organizationId"`
+	Type           GcpCredentialsType                `json:"type"`
+	Value          GoogleCostCredentialsValeuPayload `json:"value"`
+}
+
+type GoogleCostCredentialsValeuPayload struct {
+	TableId string `json:"tableid"`
+	Secret  string `json:"secret"`
+}
+
+type GcpCredentialsCreatePayload struct {
+	Name           string                     `json:"name"`
+	OrganizationId string                     `json:"organizationId"`
+	Type           GcpCredentialsType         `json:"type"`
+	Value          GcpCredentialsValuePayload `json:"value"`
+}
+
+type GcpCredentialsValuePayload struct {
+	ProjectId         string `json:"projectId"`
+	ServiceAccountKey string `json:"serviceAccountKey"`
+}
+
+const (
+	GoogleCostCredentiassType            GcpCredentialsType   = "GCP_CREDENTIALS"
+	AzureCostCredentialsType             AzureCredentialsType = "AZURE_CREDENTIALS"
+	AwsCostCredentialsType               AwsCredentialsType   = "AWS_ASSUMED_ROLE"
+	AwsAssumedRoleCredentialsType        AwsCredentialsType   = "AWS_ASSUMED_ROLE_FOR_DEPLOYMENT"
+	AwsAccessKeysCredentialsType         AwsCredentialsType   = "AWS_ACCESS_KEYS_FOR_DEPLOYMENT"
+	GcpServiceAccountCredentialsType     GcpCredentialsType   = "GCP_SERVICE_ACCOUNT_FOR_DEPLOYMENT"
+	AzureServicePrincipalCredentialsType AzureCredentialsType = "AZURE_SERVICE_PRINCIPAL_FOR_DEPLOYMENT"
+)
+
 func (self *ApiClient) CloudCredentials(id string) (Credentials, error) {
 	var credentials, err = self.CloudCredentialsList()
 	if err != nil {
