+++ client/common_models.go
@@ -1,0 +1,13 @@
+package client
+
+type User struct {
+	CreatedAt   string                 `json:"created_at"`
+	Email       string                 `json:"email"`
+	FamilyName  string                 `json:"family_name"`
+	GivenName   string                 `json:"given_name"`
+	Name        string                 `json:"name"`
+	Picture     string                 `json:"picture"`
+	UserId      string                 `json:"user_id"`
+	LastLogin   string                 `json:"last_login"`
+	AppMetadata map[string]interface{} `json:"app_metadata"`
+}
