+++ client/git_token.go
@@ -1,5 +1,17 @@
 package client
 
+type GitToken struct {
+	Id             string `json:"id"`
+	Name           string `json:"name"`
+	Value          string `json:"value"`
+	OrganizationId string `json:"organizationId"`
+}
+
+type GitTokenCreatePayload struct {
+	Name  string `json:"name"`
+	Value string `json:"value"`
+}
+
 type GitTokenCreatePayloadWith struct {
 	GitTokenCreatePayload
 	OrganizationId string `json:"organizationId"`
