+++ client/notification_project_assignment.go
@@ -2,6 +2,18 @@ package client
 
 import "fmt"
 
+type NotificationProjectAssignment struct {
+	Id                     string   `json:"id"`
+	NotificationEndpointId string   `json:"notificationEndpointId"`
+	EventNames             []string `json:"eventNames"`
+	CreatedBy              string   `json:"createdBy"`
+	CreatedByUser          User     `json:"createdByUser"`
+}
+
+type NotificationProjectAssignmentUpdatePayload struct {
+	EventNames []string `json:"eventNames"`
+}
+
 func (ac *ApiClient) NotificationProjectAssignments(projectId string) ([]NotificationProjectAssignment, error) {
 	var result []NotificationProjectAssignment
 	if err := ac.http.Get("/notifications/projects/"+projectId, nil, &result); err != nil {
