+++ client/organization.go
@@ -4,6 +4,30 @@ import (
 	"errors"
 )
 
+type Organization struct {
+	Id                                  string  `json:"id"`
+	Name                                string  `json:"name"`
+	MaxTtl                              *string `json:"maxTtl"`
+	DefaultTtl                          *string `json:"defaultTtl"`
+	DoNotReportSkippedStatusChecks      bool    `json:"doNotReportSkippedStatusChecks"`
+	DoNotConsiderMergeCommitsForPrPlans bool    `json:"doNotConsiderMergeCommitsForPrPlans"`
+	Description                         string  `json:"description"`
+	PhotoUrl                            string  `json:"photoUrl"`
+	CreatedBy                           string  `json:"createdBy"`
+	CreatedAt                           string  `json:"createdAt"`
+	UpdatedAt                           string  `json:"updatedAt"`
+	Role                                string  `json:"role"`
+	IsSelfHosted                        bool    `json:"isSelfHosted"`
+	IsSelfHostedK8s                     bool    `json:"isSelfHostedK8s"`
+}
+
+type OrganizationPolicyUpdatePayload struct {
+	MaxTtl                              *string `json:"maxTtl"`
+	DefaultTtl                          *string `json:"defaultTtl"`
+	DoNotReportSkippedStatusChecks      *bool   `json:"doNotReportSkippedStatusChecks,omitempty"`
+	DoNotConsiderMergeCommitsForPrPlans *bool   `json:"doNotConsiderMergeCommitsForPrPlans,omitempty"`
+}
+
 func (ac *ApiClient) Organization() (Organization, error) {
 	var result []Organization
 	err := ac.http.Get("/organizations", nil, &result)
