+++ client/sshkey.go
@@ -1,5 +1,23 @@
 package client
 
+type SshKey struct {
+	User           User   `json:"user"`
+	UserId         string `json:"userId"`
+	CreatedAt      string `json:"createdAt"`
+	UpdatedAt      string `json:"updatedAt"`
+	Id             string `json:"id"`
+	Name           string `json:"name"`
+	Description    string `json:"description"`
+	OrganizationId string `json:"organizationId"`
+	Value          string `json:"value"`
+}
+
+type SshKeyCreatePayload struct {
+	Name           string `json:"name"`
+	OrganizationId string `json:"organizationId"`
+	Value          string `json:"value"`
+}
+
 func (self *ApiClient) SshKeyCreate(payload SshKeyCreatePayload) (*SshKey, error) {
 	organizationId, err := self.organizationId()
 	if err != nil {
