+++ client/workflow_triggers.go
@@ -1,5 +1,13 @@
 package client
 
+type WorkflowTrigger struct {
+	Id string `json:"id"`
+}
+
+type WorkflowTriggerUpsertPayload struct {
+	DownstreamEnvironmentIds []string `json:"downstreamEnvironmentIds"`
+}
+
 func (self *ApiClient) WorkflowTrigger(environmentId string) ([]WorkflowTrigger, error) {
 	var result []WorkflowTrigger
 	err := self.http.Get("environments/"+environmentId+"/downstream", nil, &result)
